Do not delete tokens when returned MissingTokenError.
These are not returned from require_valid() as they have not successfully been refreshed, but they are also not known to be invalid so they aren't deleted.

The MissingTokenError is raised when the response is missing the 'access_token' and does not contain a rfc6749 'error' message. This can happen when the SSO servers explode and return 400-series errors.

Thanks @antihax

Version bump to v1.4.7