Reduce query joins when filtering tokens.
There was an unofficial limit of something like 22 scopes when filtering as any more exceeded the MySQL maximum tables referenced by a single join statement (61). On a 21 scope token, this change takes it from 172 joins down to 22 (don't know the table count but this probably drops it substantially). There will still be a cap of approximately 60 scopes but there's nothing I can do for that (and you should never be asking for that many scopes at once).

Additionally this ensures scopes are a unique list when filtering because if not, tokens would be incorrectly filtered out when using the require_tokens_exact method (which counts number of scopes on the token vs number of scopes requested).

And this handles space-delimited one-string scope lists like ['scope1 scope2'] because I've seen that and it costs nothing to protect against.

This should probably be several commits but it happened all at once so this is how you get it.

Thanks @markus-lassfolk